Add MapShaker / remove unused Usings
