+++ Jaeho/SnakeGame/SnakeGame/Assets/MapData/Stage_1.txt
@@ -1,0 +1,20 @@
+3
+BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                      P                                  B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
