+++ Jaeho/SnakeGame/SnakeGame/Assets/MapData/Stage_2.txt
@@ -1,0 +1,20 @@
+30
+BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
+B         B                                    B          B
+B         B                                    B          B
+B         B                                    B          B
+B         B                                    B          B
+B         B                                    B          B
+B                      P                                  B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B                                                         B
+B         B                                    B          B
+B         B                                    B          B
+B         B                                    B          B
+B         B                                    B          B
+B         B                                    B          B
+BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
