+++ Jaeho/SnakeGame/SnakeGame/Managers/ColliderManager.cs
@@ -1,10 +1,4 @@
-﻿using ConsoleGame;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Diagnostics;
 
 namespace ConsoleGame
 {
