+++ Jaeho/SnakeGame/SnakeGame/Managers/RenderManager.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SnakeGame;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Diagnostics;
@@ -40,6 +41,8 @@ namespace ConsoleGame
             {
                 renderer.Render();
             }
+
+            MapShaker.Instance.RenderShakedMap();
         }
 
         public void Release()
