+++ Jaeho/SnakeGame/SnakeGame/Managers/TimeManager.cs
@@ -14,8 +14,8 @@ namespace ConsoleGame
         }
 
         public readonly int TimeScale = 1;
-        public static int DeltaTime = 1000 / FPS;
-        public const int FPS = 10;
+        public static int MS_PER_FRAME = 1000 / FPS;
+        public const int FPS = 15;
         private Stopwatch stopwatch = new Stopwatch();
         private long _currentMs = 0;
         private long _elapsed = 0;
