+++ Jaeho/SnakeGame/SnakeGame/Objects/ScoreBoard.cs
@@ -1,0 +1,22 @@
+﻿using ConsoleGame;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SnakeGame
+{
+    public class ScoreBoard
+    {
+        ScoreBoard()
+        {
+           
+        }
+
+        public void Render()
+        {
+            
+        }
+    }
+}
